# swagger/api/swagger.yaml
---
swagger: "2.0"
info:
  description: "The API for the Pool Controller IOT project"
  version: "1.0.0"
  title: "pool-iot-api"
host: "virtserver.swaggerhub.com"
basePath: "/rhbroberg/pool-api/1.0.0"
schemes:
- "https"
consumes:
- "application/json"
produces:
- "application/json"
paths:
  /devices:
    get:
      tags:
      - "Device"
      description: "returns all configured devices"
      operationId: "getDevices"
      parameters:
      - name: "skip"
        in: "query"
        description: "number of records to skip"
        required: false
        type: "integer"
        format: "int32"
      - name: "limit"
        in: "query"
        description: "max number of records to return"
        required: false
        type: "integer"
        format: "int32"
      responses:
        200:
          description: "All the devices"
          schema:
            type: "array"
            items:
              type: "string"
              format: "uri"
              example: "http://10.0.0.225:8080"
      x-swagger-router-controller: "Device"
  /devices/{deviceId}/register:
    put:
      tags:
      - "Device"
      operationId: "registerDevice"
      parameters:
      - name: "deviceId"
        in: "path"
        required: true
        type: "string"
      responses:
        200:
          description: "successfully registered device"
          schema:
            $ref: "#/definitions/DeviceRegistrationInfo"
      x-swagger-router-controller: "Device"
  /devices/{deviceId}/{state}:
    put:
      tags:
      - "Device"
      operationId: "setDeviceState"
      parameters:
      - name: "deviceId"
        in: "path"
        required: true
        type: "string"
      - name: "state"
        in: "path"
        required: true
        type: "string"
        enum:
        - "false"
        - "true"
      responses:
        200:
          description: "set device state"
          schema:
            $ref: "#/definitions/DeviceState"
      x-swagger-router-controller: "Device"
  /devices/{deviceId}:
    get:
      tags:
      - "Device"
      operationId: "getDeviceState"
      parameters:
      - name: "deviceId"
        in: "path"
        required: true
        type: "string"
      responses:
        200:
          description: "get device state"
      x-swagger-router-controller: "Device"
  /filtration:
    get:
      tags:
      - "Filtration"
      description: "returns all registered devices"
      operationId: "getFiltrationDevices"
      parameters:
      - name: "skip"
        in: "query"
        description: "number of records to skip"
        required: false
        type: "integer"
        format: "int32"
      - name: "limit"
        in: "query"
        description: "max number of records to return"
        required: false
        type: "integer"
        format: "int32"
      responses:
        200:
          description: "All the devices"
          schema:
            type: "array"
            items:
              type: "string"
              format: "uri"
              example: "http://10.0.0.225:8080"
      x-swagger-router-controller: "Filtration"
  /filtration/{pumpId}/{state}:
    put:
      tags:
      - "Filtration"
      operationId: "setFiltrationState"
      parameters:
      - name: "pumpId"
        in: "path"
        required: true
        type: "string"
      - name: "state"
        in: "path"
        required: true
        type: "string"
        enum:
        - "false"
        - "true"
      responses:
        200:
          description: "response"
          schema:
            $ref: "#/definitions/ApiResponse"
      x-swagger-router-controller: "Filtration"
  /filtration/{pumpId}:
    get:
      tags:
      - "Filtration"
      operationId: "getFiltrationState"
      parameters:
      - name: "pumpId"
        in: "path"
        required: true
        type: "string"
      responses:
        200:
          description: "response"
          schema:
            $ref: "#/definitions/ApiResponse"
      x-swagger-router-controller: "Filtration"
  /temperature:
    get:
      tags:
      - "Environment"
      operationId: "temperatureSummary"
      parameters: []
      responses:
        200:
          description: "ok"
          schema:
            $ref: "#/definitions/TemperatureSummary"
      x-swagger-router-controller: "Environment"
  /temperature/{poolId}:
    get:
      tags:
      - "Environment"
      operationId: "getPoolTemperature"
      parameters:
      - name: "poolId"
        in: "path"
        required: true
        type: "string"
        enum:
        - "pool"
        - "spa"
      responses:
        200:
          description: "Zone temperature"
          schema:
            $ref: "#/definitions/TemperatueZoneStatus"
      x-swagger-router-controller: "Environment"
  /temperature/{poolId}/heater:
    get:
      tags:
      - "Environment"
      description: "gets the state of the heater"
      operationId: "getHeaterState"
      parameters:
      - name: "poolId"
        in: "path"
        required: true
        type: "string"
        enum:
        - "pool"
        - "spa"
      responses:
        200:
          description: "heater state"
          schema:
            $ref: "#/definitions/HeaterState"
      x-swagger-router-controller: "Environment"
  /temperature/{poolId}/heater/{state}:
    put:
      tags:
      - "Environment"
      description: "turns the heater on or off"
      operationId: "setHeaterState"
      parameters:
      - name: "poolId"
        in: "path"
        required: true
        type: "string"
        enum:
        - "pool"
        - "spa"
      - name: "state"
        in: "path"
        required: true
        type: "string"
        enum:
        - "false"
        - "true"
      responses:
        200:
          description: "Status of the operation"
          schema:
            $ref: "#/definitions/ApiResponse"
      x-swagger-router-controller: "Environment"
  /lighting:
    get:
      tags:
      - "Lighting"
      operationId: "getLights"
      parameters: []
      responses:
        200:
          description: "ok"
          schema:
            type: "array"
            items:
              type: "string"
      x-swagger-router-controller: "Lighting"
  /lighting/{lightId}:
    get:
      tags:
      - "Lighting"
      operationId: "getLightingState"
      parameters:
      - name: "lightId"
        in: "path"
        required: true
        type: "string"
        enum:
        - "underwater"
        - "yard"
      responses:
        200:
          description: "ok"
          schema:
            $ref: "#/definitions/LightingZoneStatus"
      x-swagger-router-controller: "Lighting"
  /lighting/{lightId}/{state}:
    put:
      tags:
      - "Lighting"
      operationId: "setLightingState"
      parameters:
      - name: "lightId"
        in: "path"
        required: true
        type: "string"
      - name: "state"
        in: "path"
        required: true
        type: "string"
        enum:
        - "false"
        - "true"
      responses:
        200:
          description: "Status of the operation"
          schema:
            $ref: "#/definitions/LightingZoneStatus"
      x-swagger-router-controller: "Lighting"
  /users:
    post:
      tags:
      - "User management"
      operationId: "createUser"
      parameters: []
      responses:
        200:
          description: "Status of user creation"
          schema:
            $ref: "#/definitions/User"
      x-swagger-router-controller: "UserManagement"
    x-swagger-router-controller: "UserManagement"
  /users/login:
    post:
      tags:
      - "User management"
      operationId: "loginUser"
      parameters: []
      responses:
        200:
          description: "Status of login"
          schema:
            $ref: "#/definitions/User"
      x-swagger-router-controller: "UserManagement"
  /users/me:
    get:
      tags:
      - "User management"
      operationId: "getUser"
      parameters: []
      responses:
        200:
          description: "Current user id"
          schema:
            $ref: "#/definitions/User"
      x-swagger-router-controller: "UserManagement"
  /users/me/token:
    delete:
      tags:
      - "User management"
      operationId: "logoutUser"
      parameters: []
      responses:
        200:
          description: "Logs out user"
      x-swagger-router-controller: "UserManagement"
definitions:
  User:
    type: "object"
    properties:
      id:
        type: "string"
      email:
        type: "string"
    example:
      id: "id"
      email: "email"
  LightingZone:
    type: "object"
    properties:
      id:
        type: "string"
      name:
        type: "string"
      deviceId:
        type: "integer"
        format: "int32"
      deviceType:
        type: "string"
        enum:
        - "dimmer"
        - "switch"
      zone:
        type: "string"
  LightingZoneStatus:
    type: "object"
    properties:
      id:
        type: "string"
      name:
        type: "string"
      lastUpdate:
        type: "string"
        format: "date-time"
      level:
        type: "integer"
        format: "int32"
    description: "the status of the lighting zone."
    example:
      level: 0
      lastUpdate: "2000-01-23T04:56:07.000+00:00"
      name: "name"
      id: "id"
  TemperatureSummary:
    type: "object"
    properties:
      zones:
        type: "array"
        items:
          $ref: "#/definitions/TemperatureZone"
      zoneStatus:
        type: "array"
        items:
          $ref: "#/definitions/TemperatueZoneStatus"
    description: "ok"
    example:
      zoneStatus:
      - name: "name"
        id: "id"
        units: "fahrenheit"
        value: 5.962133916683182377482808078639209270477294921875
        timestamp: "2000-01-23T04:56:07.000+00:00"
      - name: "name"
        id: "id"
        units: "fahrenheit"
        value: 5.962133916683182377482808078639209270477294921875
        timestamp: "2000-01-23T04:56:07.000+00:00"
      zones:
      - inputPosition: 6
        outputPosition: 1
        zone: "zone"
        name: "name"
        id: 0
      - inputPosition: 6
        outputPosition: 1
        zone: "zone"
        name: "name"
        id: 0
  TemperatureZone:
    type: "object"
    required:
    - "id"
    - "name"
    properties:
      id:
        type: "integer"
        format: "int32"
        description: "the unique identifier for the zone"
      name:
        type: "string"
      inputPosition:
        type: "integer"
        format: "int32"
      outputPosition:
        type: "integer"
        format: "int32"
      zone:
        type: "string"
    description: "a single temperature zone"
    example:
      inputPosition: 6
      outputPosition: 1
      zone: "zone"
      name: "name"
      id: 0
  TemperatueZoneStatus:
    type: "object"
    required:
    - "id"
    - "timestamp"
    - "value"
    properties:
      id:
        type: "string"
        description: "the unique identifier for the zone"
      name:
        type: "string"
        description: "the name of the zone"
      value:
        type: "number"
        format: "double"
        description: "the temperature in the zone"
      units:
        type: "string"
        description: "the temperature units"
        enum:
        - "celcius"
        - "fahrenheit"
        default: "fahrenheit"
      timestamp:
        type: "string"
        format: "date-time"
        description: "the timestamp when the temperature was measured"
    description: "status of a single zone"
    example:
      name: "name"
      id: "id"
      units: "fahrenheit"
      value: 5.962133916683182377482808078639209270477294921875
      timestamp: "2000-01-23T04:56:07.000+00:00"
  ApiResponse:
    type: "object"
    properties:
      code:
        type: "integer"
        format: "int32"
      message:
        type: "string"
        example: "everything is ok"
    example:
      code: 0
      message: "everything is ok"
  HeaterState:
    type: "object"
    properties:
      id:
        type: "string"
      state:
        type: "string"
    example:
      id: "id"
      state: "state"
  DeviceState:
    type: "object"
    properties:
      id:
        type: "string"
      name:
        type: "string"
      lastUpdate:
        type: "string"
        format: "date-time"
      level:
        type: "integer"
        format: "int32"
    example:
      level: 0
      lastUpdate: "2000-01-23T04:56:07.000+00:00"
      name: "name"
      id: "id"
  DeviceRegistrationInfo:
    type: "object"
    properties:
      uri:
        type: "string"
        format: "uri"
        example: "http://10.0.0.220:8080"
      id:
        type: "string"
        format: "uuid"
        example: "0729a580-2240-11e6-9eb5-0002a5d5c51b"
    example:
      id: "0729a580-2240-11e6-9eb5-0002a5d5c51b"
      uri: "http://10.0.0.220:8080"